+++ src/main/java/com/retailmax/notifications/DataLoader.java
@@ -1,4 +1,4 @@
-package com.retailmax.notificaciones;
+package com.retailmax.notifications;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
