+++ src/main/java/com/retailmax/notifications/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.retailmax.notificaciones.config;
+package com.retailmax.notifications.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
