+++ src/main/java/com/retailmax/notifications/model/PreferenciaDeNotificacion.java
