+++ src/main/java/com/retailmax/notifications/model/Usuario.java
