+++ src/main/java/com/retailmax/notifications/service/PromocionService.java
@@ -1,4 +1,4 @@
-package com.retailmax.notificaciones.service;
+package com.retailmax.notifications.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
