+++ src/test/java/com/retailmax/notificaciones/service/PromocionServiceTest.java
@@ -1,4 +1,4 @@
-package com.retailmax.notificaciones.service;
+package com.retailmax.notifications.service;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
